db setup for attendance
