+++ home/templates/home/classroom_detail.html
@@ -156,6 +156,11 @@ function revertASF() {
     <p style="padding-left: 20px;"><button id="add_stus_button" class="button" onclick="addStudentsForm()">Add Students</button></p>
 
   <hr>
+  <br>
+  <button class="button" >Take Attendance Today</button></p><br>
+
+  <button class="button" >All Attendance</button></p><br>
+
   </div>
