+++ staff/staff_views_helper.py
@@ -1,12 +1,14 @@
 from django.contrib.auth.models import User as DjangoUser
 from django.contrib.auth.models import Group
-from home.models.salesforce import ClassEnrollment
-from home.models.models import UserProfile, Classroom
+from home.models.salesforce import ClassEnrollment, Contact, ClassOffering
+from home.models.models import UserProfile, Classroom, Attendance, Session
 from django.core.mail import send_mail
 from django.template.loader import render_to_string
 from django.utils.html import strip_tags
 from django.conf import settings
 from django.contrib import messages
+from home.forms import AddVolunteersForm, AddStudentsForm
+from datetime import timedelta
 
 
 def create_user_with_profile(form, random_password):
@@ -194,3 +196,146 @@ def add_volunteers_to_volunteer_dict(classroom):
             volunteer_user.last_name,
         )
     return volunteer_dict
+
+
+def change_classroom_teacher(request):
+    teacher_contact = get_contact_by_user_id(request.POST["former_teacher"])
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    new_teacher = get_contact_by_user_id(request.POST["teacher"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    classroom.teacher_id = request.POST["teacher"]
+    classroom.save()
+    remove_enrollment(teacher_contact, class_offering)
+    ClassEnrollment.objects.get_or_create(
+        created_by=class_offering.created_by,
+        contact=new_teacher,
+        status="Enrolled",
+        class_offering=class_offering,
+    )
+    class_offering.instructor = new_teacher
+    class_offering.save()
+
+
+def change_classroom_ta(request):
+    ta_contact = get_contact_by_user_id(request.POST["former_ta"])
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    new_ta = get_contact_by_user_id(request.POST["teacher"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    classroom.teacher_assistant_id = request.POST["teacher"]
+    classroom.save()
+    remove_enrollment(ta_contact, class_offering)
+    ClassEnrollment.objects.get_or_create(
+        created_by=class_offering.created_by,
+        contact=new_ta,
+        status="Enrolled",
+        class_offering=class_offering,
+    )
+
+
+def remove_volunteer(request):
+    vol_contact = get_contact_by_user_id(request.POST["former_vol"])
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    classroom.volunteers.remove(request.POST["former_vol"])
+    remove_enrollment(vol_contact, class_offering)
+
+
+def remove_student(request):
+    student_contact = get_contact_by_user_id(request.POST["former_student"])
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    classroom.students.remove(request.POST["former_student"])
+    remove_enrollment(student_contact, class_offering)
+
+
+def add_volunteers(request):
+    form = AddVolunteersForm(request.POST)
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    if form.is_valid():
+        for volunteer in form.cleaned_data.get("volunteers"):
+            vol_contact = get_contact_by_user_id(volunteer)
+            ClassEnrollment.objects.create(
+                created_by=class_offering.created_by,
+                contact=vol_contact,
+                status="Enrolled",
+                class_offering=class_offering
+            )
+            classroom.volunteers.add(volunteer)
+
+
+def add_students(request):
+    form = AddStudentsForm(request.POST)
+    class_offering = get_class_offering_by_id(request.POST["course_id"])
+    classroom = Classroom.objects.get(id=request.POST["course_id"])
+    if form.is_valid():
+        for student in form.cleaned_data.get("students"):
+            student_contact = get_contact_by_user_id(student)
+            ClassEnrollment.objects.create(
+                created_by=class_offering.created_by,
+                contact=student_contact,
+                status="Enrolled",
+                class_offering=class_offering
+            )
+            classroom.students.add(student)
+
+
+def remove_enrollment(contact, class_offering):
+    old_enrollment = ClassEnrollment.objects.get(contact=contact, status="Enrolled", class_offering=class_offering)
+    created_by = old_enrollment.created_by
+    old_enrollment.delete()
+    return created_by
+
+
+def get_contact_by_user_id(id):
+    return Contact.objects.get(client_id=UserProfile.objects.get(user_id=id).salesforce_id)
+
+
+def get_class_offering_by_id(id):
+    course_name = get_course_name_by_id(id)
+    return ClassOffering.objects.get(name=course_name)
+
+
+def get_course_name_by_id(id):
+    return Classroom.objects.get(id=id).course
+
+
+def sync_attendance_with_salesforce_class_offerings():
+    """
+    This method is fluid; in its current state it will duplicate entries
+    in your postgres attendance database if you've already got attendance data!
+    Only use if you don't have attendance data but do have Classroom
+    data.
+    """
+    for classoffering in ClassOffering.objects.all():
+        dates = class_offering_meeting_dates(classoffering)
+        classroom = Classroom.objects.get(course=classoffering.name)
+        for day in dates:
+            session = Session.objects.create()
+            for student in classroom.students.all():
+                Attendance.objects.create(
+                    student_id=student.id,
+                    session_id=session.id,
+                    classroom_id=classroom.id,
+                    date=day
+                )
+
+
+def class_offering_meeting_dates(class_offering):
+    int_days = get_integer_days(class_offering)
+    class_range = class_offering.end_date - class_offering.start_date
+    dates = []
+    for i in range(class_range.days + 1):
+        the_date = class_offering.start_date + timedelta(days=i)
+        if the_date.weekday() in int_days:
+            dates.append(the_date)
+    return dates
+
+
+def get_integer_days(class_offering):
+    if class_offering.meeting_days == "M-F":
+        return [0, 1, 2, 3, 4]
+    elif class_offering.meeting_days == "M/W":
+        return [0, 2]
+    elif class_offering.meeting_days == "T/R":
+        return [1, 3]
